add analyzer

+++ telemetry/collector/src/main/java/ru/yandex/practicum/telemetry/collector/gRPC/builders/hub/ScenarioAddedBuilder.java
@@ -78,7 +78,7 @@ public class ScenarioAddedBuilder extends BaseHubBuilder {
                                     case UNRECOGNIZED -> null;
                                 }
                         )
-                        .setValue(da.getValue())
+                        .setValue((int) da.getValue())
                         .build())
                 .toList();
     }
